create logs and csv

+++ Cloud/code/cloud_server.py
@@ -1,6 +1,9 @@
 # Importing the FastAPI framework and the CloudMain class from another module
-import time
+import csv
 import datetime
+import os.path
+import time
+
 from fastapi import Body
 from fastapi import FastAPI
 
@@ -17,6 +20,19 @@ def create_logs(txt_file, txt_log):
     txt_file.write(f"{date_string} cloud_server   | {txt_log}\n")
 
 
+def create_data(t_id, desc, start, end):
+    row = [t_id, desc, start, end, end - start]
+    return row
+
+
+def save_csv(csv_file, csv_data):
+    # open a new CSV file in write mode
+    writer = csv.writer(csv_file)
+
+    # write each row to the CSV file
+    writer.writerow(csv_data)
+
+
 # Defining a route for checking the connection to the Cloud Server
 @app.get("/check/connection/cloud")
 async def check_connection():
@@ -27,7 +43,16 @@ async def check_connection():
 @app.post("/cloud/information")
 async def cloud1(road_info: dict = Body(...)):
     start_time = time.time()  # get the current time in seconds
-    txt_file = open('/data/cloud_logs.txt', 'w')
+
+    if os.path.exists('/data/cloud_logs.txt'):
+        txt_file = open('/data/cloud_logs.txt', 'a')
+        csv_file = open('/data/cloud_times.csv', "a", newline="")
+    else:
+        txt_file = open('/data/cloud_logs.txt', 'w')
+        csv_file = open('/data/cloud_times.csv', "w", newline="")
+        header = ["ID", "Description", "Start Time", "End Time", "Total Time"]
+        (csv.writer(csv_file)).writerow(header)
+
     # Printing the received information for debugging purposes
     print(f"road_info: {road_info}")
     create_logs(txt_file, f"road_info: {road_info}")
@@ -42,6 +67,7 @@ async def cloud1(road_info: dict = Body(...)):
     print(f"Elapsed time for cloud: {elapsed_time} seconds")
     str_time = f"Elapsed time for cloud: {elapsed_time} seconds"
     create_logs(txt_file, str_time)
+    save_csv(csv_file, create_data(0, "StartEndTime", start_time, end_time))
 
     # Returning the response from the 'road_check' method
     return response
